Add metadata to movement_libras
Google collab notebook:  https://colab.research.google.com/drive/1opNT7NFkOS0KxIcNhg_Jl3Uka-4QKtQI?usp=sharing

Note: The openML description was very detailed, and I did not think it was necessary to include all of that detail in the metadata here. I am not sure if what I have written is enough or too much; if it is too little, I can go back and be more specific and verbose in the general description.