Adjusted quotes
Removed all instances of two single quotes next to each other ('') for ease of reading and coding. This was initially in place to reflect that the source data contains values with single quotes around it, but for someone trying to use this metadata in their code, the '' might be read as an empty string.