Replacing `pandas-profiling` (deprecated) with `ydata-profiling`
the previously known pandas-profiling is now part of a bigger
project and is decoupling from the idea that is intended to be
used only with dataframes.

The name of the package has changed, and the last version of
`pandas-profiling` was released more than a year ago.

The github workflow for profiling new datasets is not working
as it should, due to deprecated dependences.